chore(quality): port lyra quality gates + split cli into subpackage (#1)
Ports lyra's code-quality guards — these should arrive via /init/stack.yml
but currently require per-repo setup (tracked in memory for dev-core follow-up):

- tools/check_file_length.sh — max 300 LOC per src/**/*.py (+ exemptions file)
- tools/check_folder_size.sh — max 12 .py files per folder (+ exemptions file)
- import-linter — architectural layers via .importlinter (3 forbidden contracts)
- .pre-commit-config.yaml — adds check-file-length, check-folder-size, import-layers

Split src/llmcli/cli.py (352 LOC) into src/llmcli/cli/ subpackage:
- _app.py (typer.Typer + console singletons)
- catalog.py (list, pull) — 83 LOC
- lifecycle.py (serve, stop, status) — 117 LOC
- swap.py (swap) — 35 LOC
- chat.py (chat) — 49 LOC
- proxy.py (register-proxy) — 83 LOC
- __init__.py re-exports patch targets; lazy-import pattern preserves
  all `patch("llmcli.cli.X")` mock targets — zero test churn

All 199 tests still pass. All quality gates green.
