+++ .pre-commit-config.yaml
@@ -11,6 +11,21 @@ repos:
         entry: uv run ruff check --select=PYI .
         language: system
         pass_filenames: false
+      - id: check-file-length
+        name: check-file-length
+        entry: tools/check_file_length.sh
+        language: system
+        pass_filenames: false
+      - id: check-folder-size
+        name: check-folder-size
+        entry: tools/check_folder_size.sh
+        language: system
+        pass_filenames: false
+      - id: import-layers
+        name: import-layers
+        entry: uv run lint-imports
+        language: system
+        pass_filenames: false
       - id: gitleaks
         name: gitleaks secret scan
         entry: gitleaks detect --source . --no-git
