+++ pyproject.toml
@@ -46,6 +46,7 @@ dev = [
     "pytest-timeout>=0.5",
     "pre-commit>=4.5.1",
     "pip-licenses>=5.5.1",
+    "import-linter>=2.1",
 ]
 vllm = [
     "vllm>=0.6.0",
