+++ src/llmcli/cli/_app.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+import typer
+from rich.console import Console
+
+app = typer.Typer(add_completion=False, help="llmCLI — local LLM serving")
+console = Console()
+err_console = Console(stderr=True)
