+++ src/llmcli/cli/proxy.py
@@ -1,0 +1,83 @@
+from __future__ import annotations
+
+import socket
+import subprocess
+from pathlib import Path
+from typing import Optional
+
+import typer
+
+from llmcli.cli._app import app, console, err_console
+
+
+# ---------------------------------------------------------------------------
+# register-proxy
+# ---------------------------------------------------------------------------
+
+
+@app.command(name="register-proxy")
+def register_proxy(
+    config_path: Optional[str] = typer.Option(
+        None,
+        "--config",
+        envvar="LITELLM_CONFIG_PATH",
+        help="Path to LiteLLM config.yaml (default: ~/.litellm/config.yaml)",
+    ),
+) -> None:
+    """Refresh the llmCLI block in ~/.litellm/config.yaml and reload."""
+    import llmcli.cli as _cli
+
+    # 1. Load catalog
+    catalog = _cli.config.load()
+
+    # 2. Determine host
+    hostname = socket.gethostname()
+
+    # 3. Resolve config path: --config flag > env var > default
+    resolved_path = (
+        Path(config_path) if config_path else Path.home() / ".litellm" / "config.yaml"
+    )
+
+    # 4. Friendly error when parent directory doesn't exist or is not writable
+    if not resolved_path.parent.exists():
+        err_console.print(
+            f"[red]Config directory does not exist: {resolved_path.parent}[/red]\n"
+            f"Create it first:  mkdir -p {resolved_path.parent}"
+        )
+        raise typer.Exit(code=1)
+
+    # 5. Build and write block
+    block = _cli.build_block(catalog, catalog.host.public_base_url)
+    try:
+        _cli.write_block(block, resolved_path)
+    except PermissionError as exc:
+        err_console.print(
+            f"[red]Permission denied writing to {resolved_path}: {exc}[/red]\n"
+            f"Check file permissions or run with appropriate privileges."
+        )
+        raise typer.Exit(code=1)
+    except OSError as exc:
+        err_console.print(f"[red]Failed to write config: {exc}[/red]")
+        raise typer.Exit(code=1)
+
+    model_count = len(catalog.models)
+
+    # 6. Reload proxy — warn on failure, don't fail the command
+    try:
+        _cli.reload_proxy()
+        reload_status = "[green]reloaded[/green]"
+    except (subprocess.CalledProcessError, FileNotFoundError, OSError) as exc:
+        reload_status = f"[yellow]reload failed (write succeeded): {exc}[/yellow]"
+        err_console.print(
+            f"[yellow]Warning: proxy reload failed — {exc}[/yellow]\n"
+            "The config file was updated successfully. Reload the proxy manually."
+        )
+
+    # 7. Confirmation output
+    console.print(
+        f"[green]LiteLLM proxy config updated.[/green] "
+        f"host=[cyan]{hostname}[/cyan] "
+        f"path=[cyan]{resolved_path}[/cyan] "
+        f"models=[bold]{model_count}[/bold] "
+        f"reload={reload_status}"
+    )
