+++ tools/file_exemptions.txt
@@ -1,0 +1,3 @@
+# File size exemptions (≤300 lines per file)
+# Each entry must have a tracking issue.
+# Format: <path>  # <lines> lines — <issue> <description>
